% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boss.R
\name{boss_choose}
\alias{boss_choose}
\title{boss_choose}
\usage{
boss_choose(x)
}
\arguments{
\item{x}{}
}
\value{
character
}
\description{
boss_choose
}
\examples{
boss_choose(c("Harry","Sally","Bob","Frieda"))
}
