% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{boss_slack_choose}
\alias{boss_slack_choose}
\title{boss_slack_choose}
\usage{
boss_slack_choose(x, channel)
}
\arguments{
\item{x}{}
}
\description{
boss_slack_choose
}
\examples{
boss_slack_choose(c("Harry","Sally","Bob","Frieda"), channel = "general")
}
