% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{boss_slack_week}
\alias{boss_slack_week}
\title{boss_slack_week}
\usage{
boss_slack_week(x, channel)
}
\arguments{
\item{x}{}
}
\description{
boss_slack_week
}
\examples{
boss_slack_week(c("Harry","Sally","Bob","Frieda"))
}
