% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boss.R
\name{boss_week}
\alias{boss_week}
\title{boss_week}
\usage{
boss_week(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
data.frame of the work-week with the boss for each day
}
\description{
boss_week
}
\examples{
boss_week(c("Harry","Sally","Bob","Frieda"))
}
